Create flag for FAHC improvements; add placeholder page
When the HUD_TOOL_IMPROVEMENTS flag is enabled, show the new template
which is stored in jinja2/housing_counselor/improved_index.html.
When the flag is not enabled, show the existing find-a-housing-counselor
page. The flag is disabled by default.

When the new page is enabled and a zipcode parameter is included in the
GET request, show the context values returned in the response, to aid in
front-end development in future commits.