Interpret ambiguous to_date filter as end of time period
Previously, when a user would enter a non-specific date in the
"to" field of a filterable list, (e.g. "2016" or "5-2017"), it would be
interpreted as the beginning of that time period (e.g. 1/1/2016 and
5-1-2017, respectively).
For better user experience we should use the end of that date range
instead (e.g. 12/31/2016 and 5/31/2017, respectively). This commit adds
methods to calculate what the entered date should be interpreted as.
The following commit inserts that value into the form.