Show the date values interpreted from user input
When displaying the Filterable List form, show the dates that we've
interpreted from the user's input, rather than simply showing what
they entered. This helps when the user input is ambiguous or incomplete.

When they input "2017", that translates to actual search values of
 - "1/1/2017" in the "from" field
 - "12/31/2017" in the "to" field

When they put a value in the "from" but not the "to", use today's date
as the "to" value

When they put a value in the "to" but not the "from", use the
publication date of the earliest possible result of this filter as the
"from" value.