add auto-generated streamfield documentation
This commit adds a new page of reference documentation to the site that
describes our use of Wagtail StreamFields. It lists all custom
StreamField blocks that can be added to a page (either directly or as a
child of another block that can) and all pages/models that can include
those blocks.

Due to the large number of blocks and fields, I created a script to
automatically generate the document. I've committed the script as a
Django management command and included documentation on how to run it.

As a follow up to this PR I'd like to add this script to tox (and add a
check for up-to-date documentation to Travis), but am committing this
first to solicit review on the content/format.

See http://localhost:8000/streamfield-blocks/ when running the local
documentation server to review the content.