update documentation .gitignore
This commit removes a deprecated .gitignore entry for docs/scripts
(because there are no scripts in this location) but adds one for site/,
the target directory for the MkDocs build.

Because Travis builds the documentation from the source Markdown files,
and creates its own site/ directory which then gets pushed to gh-pages,
local changes to site/ should never be committed. Adding this directory
to the .gitignore means that local documentation builds won't show up in
a git diff.