bumps django version to 1.11.18
On January 4th Django released version 1.11.18 which is a security
release that fixes a content spoofing possibility:

https://docs.djangoproject.com/en/2.1/releases/1.11.18/

We aren't currently vulnerable to this as far as I can tell because we
use a custom 404 page, but we should still upgrade. We're currently on
version 1.11.15.

All unit tests pass properly with this version. All active satellite
apps are now defined with a requirement in setup.py like
'Django>=1.11,<1.12' and so no changes should be required in them.