move transition utilities code into core
Currently we have a second-level directory `cfgov/transition_utilities`
which contains a single file that supports conditional URLs. As the
second-level under `cfgov` is mainly Django apps, there's no good reason
to maintain this folder just for one module.

The "transition" referred to in the name of this folder had to do with
old Sheer sites (I think), which we no longer support:

https://github.com/cfpb/cfgov-refresh/pull/1351

All unit tests pass with this change.