Prevent ValueError when date entered is < 1900
Previously, when a user entered a date earlier than 1/1/1900 in the Date
Range fields on a Filterable List page, they would receive a 500 error.
This happened because the set_interpreted_date_values method in this
file uses strftime to format date input for display. strftime can only
handle dates whose year is 1900 or later.

Now, when the user enters a date earlier than 1900, they see a
validation error that says "Please enter a date 1/1/1900 or later."