Show a customizable message when there are no posts
Previously, when a Filterable List page had no pages to filter (e.g. the
Events page right now), the page would show the filter controls, which
didn't make sense, and which had a broken Topic multi-select.

Now, a Wagtail user can specify a message that will show up instead of
the filter controls when there are no pages to show. The "No posts
message" field is the <h4> header and the "No posts explanation" is the
<p> body of the notification.

When there are pages to filter, nothing changes.