enforce only one filterable list
This change modifies the behavior of BrowseFilterablePage and
SublandingFilterablePage to ensure that at most one FilterControls
block is entered into the page's content StreamField.

We only ever want at most one of these blocks on a page. The current
FilterableListMixin for pages only supports one block of this type on a
page:

https://github.com/cfpb/cfgov-refresh/blob/771cbe30c587daaeb7c544bb3e3cac390255787a/cfgov/v1/util/filterable_list.py#L9

This logic could probably be expanded to other pages and blocks.