Point to correct S3 bucket for Housing Counselor files
The JSON and PDF files that power the Find a Housing Counselor tool are
stored on files.consumerfinance.gov for all instances of cf.gov,
regardless of where other S3 files (e.g. Wagtail uploads, etc) are
stored for that instance.

So, the HOUSING_COUNSELOR_S3_PATCH_TEMPLATE should not depend on the
AWS_STORAGE_BUCKET_NAME setting, as it did before. This commit updates
that setting to always point to files.consumerfinance.gov, and removes
the s3utils.py utility, since it was not called by any other code.