Allows page type-specific post preview rendering.
This change proposes an alternative to #4772 and #4790.

This change makes it possible to disable the rendering of post preview
tags and dates when rendered as part of a filterable list. This is now
possible via code in the v1.atomic_elements.organisms.FilterableList
StreamField block, and is dependent on the "page_type" set on the FL,
which is an optional setting that a page editor can define.

This change adds a new page type of "cfpb-researchers" and disables
rendering of post preview dates for it. It also disables rendering of
post preview tags for the "foia-freq-req-record" page type.

A side effect of this change is to add a proper Wagtail StreamField
block template for the PostPreviewSnapshot block type.