Removes deprecated HTMLBlock
This commit removes the deprecated v1.atomic_elements.organisms
HTMLBlock StreamField block. This block was created as part of
prototyping for the Consumer Credit Trends chart, but was never used in
production (instead replaced by ChartBlock).

The block works by making an HTTP request to an arbitrary HTML source at
render time. It is being removed because it is not being used in an
effort to clean up the code and avoid confusion with Wagtail's built-in
RawHTMLBlock, which we do use in several places.

This change also removes a unit test for this block that makes an HTTP
request to GitHub every time the test is run.