Change page model to handle editing of Ask entries (#4818)
This is the first step in migrating Ask content to a simpler editing
structure. It exposes content on the AnswerPage and allows version
comparisons.

Metadata will be handled in subsequent steps.

Notes
* The 'tags' and 'tags_es' properties on the Answer model, which are
used in indexing, are being changed to 'clean_tags' and 'clean_tags_es'
to avoid a future collision with the AnswerPage's 'tags' field.