Deprecate non-Wagtail mega menu
This change removes the WAGTAIL_MENU feature flag and removes the
template for the non-Wagtail mega menu.

The Wagtail mega menu has been turned on in production since August 2018
(over six months), but the cfgov-refresh repository still defaults to
using the non-Wagtail version unless the WAGTAIL_MENU flag is set. It
happens to be set in production database dumps, but not when starting
from an empty database.

This change removes the WAGTAIL_MENU flag and the non-Wagtail menu
template.