Add ability to log DB queries to console
This commit adds the ability to (optionally) log Django database queries
to the console by defining a new ENABLE_SQL_LOGGING environment variable.

This is useful when trying to debug database interactions, and works
against the Django server, interactive shell, and management commands.

The documentation on developer tips has been updated to include this
ability.