Remove custom password reset email template
As a follow-up to #4834 and #4870, this removes our custom password
reset email. Removing this will let us fallback to Wagtail's built-in
template, which contains the same content:

https://github.com/wagtail/wagtail/blob/v1.13.4/wagtail/wagtailadmin/templates/wagtailadmin/account/password_reset/email.txt

We do lose our custom "Thanks for using our site!" message but this
seems a reasonable tradeoff to remove one piece of customization.