Rename format variable in housing counselor view
Python contains a built-in "format" function:

https://docs.python.org/3/library/functions.html#format

>>> format(123, "06d")
'000123'

The use of "format" as a variable in this method confuses our static
code analysis tool (Checkmarx) because of this. This commit changes that
variable to something more specific, "file_format".

To test, visit a housing counselor URL locally and confirm that the data
continues to load properly:

http://localhost:8000/find-a-housing-counselor/?zipcode=22203

Also confirm that the "Save list as PDF" link continues to work
properly.