Add HMDA Explore page temporarily, with feature flag
Add a static page that duplicates what is currently served at
/data-research/hmda/explore. In the near future, we will be removing the
static site that is currently served by Apache at /data-research/hmda.
This is the only page from that site we will still need.

With the HMDA_LEGACY_REVIEW flag, it will be served at
/data-research/hmda-new/explore, so that stakeholders can review it on
content.cf.gov. When we're ready to publish the page, (and when we've
updated the Apache settings not to redirect traffic for that url), we'll
switch to the HMDA_LEGACY_PUBLISH flag, and it will serve the page at
/data-research/hmda/explore.