Fix dump_as_csv behavior in Python 3
The unicodecsv package that we use to dump CSVs requires that the
underlying stream be opened to write binary data. It works fine if you
don't do that in Python 2, but fails in Python 3.

This change improves the existing dump_as_csv unit test to make it
actually test that something gets dumped to a CSV. This test would fail
in Python 3 if not for the change from 'w' to 'wb' in the open() call.