Fix beta banner by marking notification safe
This commit fixes a problem with the beta notification banner where the
link to the real site shows up as HTML tags instead of as a proper link.
This was caused by the change in #4879 that removed a |safe tag from the
notification module.

To fix this, we just apply |safe directly to the string being passed to
the notification.

To test, run this locally and confirm that the banner looks correct on
this branch but wrong on master:

$ DEPLOY_ENVIRONMENT=beta ./runserver.sh