Fix safe HTML in wellbeing notification
This commit fixes a problem with a notification module on the Financial
Wellbeing Results page. When accessing the results page directly,
without results, the warning message currently shows raw HTML due to a
change in #4879. The message needs to be marked with `| safe`.

This is a similar fix to #4897.

To test, visit
http://localhost:8000/consumer-tools/financial-well-being/results/
locally.