Run full tox tests on Travis
This change modifies how the Python tests are run on Travis so that the
"slow" version is run instead of the "fast" version. The "slow" version
runs the full set of Django migrations.

It also modifies our Travis configuration to use tox-travis, which
allows us to simplify the .travis.yml a bit. This PR also removes the
separate missing-migrations environment in favor of moving the
migrations check into the standard unittest environment, but only having
it run on Travis using Python 2.7. This is accomplished via a bit of
environment variable trickery.