Add translations for portal topic and portal headings
Topic and portal headings are used in common by English and Spanish
answer pages, and each has an English and Spanish heading that can be set
by an editor when creating a new tag.

We add the headings to our translation `django.po` files so that templates
can call a topic tag and get the right translation.

This also enables translations on Ask's TagResultsPage to allow those
pages to use a single template when Spanish Ask is migrated.