Add a portal_topic field to SublandingPage
The portal_topic relation will allow an editor to assign a portal page
to a topic, which will define portal pages in Wagtail and avoid having
to have a hard-coded list. Portal pages use the SublandingPage model and
have can have SublandingPage siblings that are not portals.

The relation will also enable creation of links, labels and nav items
where needed, in both English and Spanish.