Update wagtail-inventory to 0.7
This change upgrades our version of wagtail-inventory from 0.5.1 to 0.7.
The later versions support more modern versions of Django (2.1, 2.2) and
Wagtail (2.3, 2.4, 2.5).

This doesn't impact us as we're not on those versions yet but it's nice
to keep up to date.

To test, run a local server and visit
http://localhost:8000/admin/inventory/ to verify that the inventory
search still works.