Remove dependency on `INTERNAL_IPS` for DDT
This change will make the Django Debug Toolbar show when running with the `local.py` settings when `ENABLE_DEBUG_TOOLBAR=True`. It makes the decision to show the toolbar based on `settings.DEBUG` rather than an IP address. This should fix issues of having to set the Docker gateway IP in .env in order to see the toolbar.