Fix bug in portal search-page routable URL handling
It's possible to manually enter a bad portal category in a portal see-all URL
and produce a server error. These URLs are normally generated by our code,
but a manually introuced typo should be caught and return a 404,
rather than cause a server error.

We already have a dictionary of category slugs available on the page,
so the fix is easy: Check any incoming category slug against the
white list before proceeding, and raise 404 if there's a mismatch.

Testing
Try this URL locally before and after loading this branch:

<http://localhost:8000/consumer-tools/debt-collection/answers/know-your-righ/>

Before: You should get a server error.
After: You should get a 404 page.