More minor improvements to initial_data script
This followup to #5039 makes additional minor changes to the
initial_data script. We want to be able to run this against any cf.gov
environment, including production, so this change makes superuser
creation optional. A superuser will only be created if the
DJANGO_ADMIN_USERNAME and DJANGO_ADMIN_PASSWORD environment variables
are set (as we suggest in our .env_SAMPLE file).

This commit also updates the environment variable used to define the
sharing hostname, from DJANGO_STAGING_HOSTNAME to a more descriptive
WAGTAIL_SHARING_HOSTNAME.

Additionally, this change slightly modifies when the legacy Wagtail
"hello world" page is deleted. This should only be necessary once, the
single time that we need to create a CFGov homepage.