Add a display_order field to PortalCategory to control nav sequence.
We want category nav choices to follow this order, to reflec the order
used on portal topic pages:
* Basics
* Key terms
* Common issues
* Know your rights
* How-to guides

To give editors control over the order, in case categories change
in the future, we're adding a `display_order` field that allows an editor
to assign a sort number to a category.

The migration of the new field includes a data migration that assigns
display_order to the initial set of categories, according to the
order above.