Recommend creation of cfpb Postgres schema
Our live cf.gov environments connect to AWS RDS Postgres instances that
are configured with a "cfpb" user and also a "cfpb" schema. The
existence of this schema means that all new database tables (e.g. ones
created by Django) are created in this cfpb schema.

Production database dumps thus refer to tables named "cfpb.something",
and include a statement to create the "cfpb" schema. When running the
local refresh-data.sh script, that loads one of these dumps, tables are
created in this schema.

On the other hand, if you run the initial-data.sh script, that attemps
to create tables locally without using a production dump, if you don't
have a "cfpb" schema, the tables are instead created in the "public"
schema. See https://www.postgresql.org/docs/10/ddl-schemas.html for
documentation on why this happens.

We want local development to be as consistent with production as
possible, and we also want use of new cf.gov environments using
initial-data.sh to work consistently with ones using production dumps.

For this reason, we should recommend and document the creation of a
"cfpb" schema as part of our Postgres setup. This change adds that line
to the documentation.