Stop redirects for Ask search
Our ask-search.html template makes a search request without a trailing slash,
which invokes Django's [APPEND_SLASH](https://docs.djangoproject.com/en/dev/ref/settings/#append-slash) default setting.
The result is a redirect for every base search request.

We just need to add a slash to the template to stop the 301s.

Note
The URL is translated for Spanish pages, but we don't need to touch the django.po file because
there was already a translation for the search URL that included the slash.