Pin django-js-asset package to match TDP
The TDP satellite app includes pinned package dependencies for
django-mptt==0.9.0 and django-js-asset==1.1.0. Currently cfgov-refresh
only includes the pinned dependency on django-mptt.

The django-mptt dependency includes an unpinned [0] dependency on
django-js-asset. There have been several releases of django-js-asset
since the 1.1.0 version [1]. This means that, depending on how the
cfgov-refresh and TDP dependencies are installed, you might end up with
a later version of django-js-asset, that doesn't match what is specified
in TDP.

For this reason, this PR fixes the django-js-asset here in
cfgov-refresh, to ensure that it always matches what is specified in
TDP.

[0] https://github.com/django-mptt/django-mptt/blob/615cd547625bef2ce86c46d7e071787f194ac924/setup.py#L20
[1] https://pypi.org/project/django-js-asset/#history