Remove references to the college-costs satellite repo
A few places in our documentation, including an example for running
a satellite app in our Docker setup, refer to the college-costs repository,
which has been folded into the cfgov-refresh repo.

One reference, in `code.json`, is left as-is, because that file includes
a number of retired repos and is a CFPB repo history of sorts.