Add quarterly FOIA reports
These quarterly FOIA report archives are already published on the
website, and listed on a Wagtail page at

https://www.consumerfinance.gov/foia-requests/foia-electronic-reading-room/cfpb-foia-quarterly-reports/

Currently these come from an internal repository and are copied into the
build as part of our deployment process. We are moving these to
cfgov-refresh to simplify things.

This is the first of multiple steps related to these files. After this
PR, they will continue to be available at /foia/quarterly, as they are
now, for example

https://www.consumerfinance.gov/foia/quarterly/CFPB-2018-Q4.zip

but now will also be available under /static at

https://www.consumerfinance.gov/static/foia/quarterly/CFPB-2018-Q4.zip

(you can verify this locally with this PR using localhost.)

A subsequent PR will modify the existing Apache alias to point to these
files.