Add Docker-based drama-free-django build
This commit adds the ability to execute the cf.gov drama-free-django [0]
build using a CentOS 6 Docker container. This creates a DFD build that
can be successfully run using Python 2 on a CentOS 6 system.

This makes public some of our build process which is currently
configured via an internal Jenkins repository.

To build and run the container, execute this command from the root of
the cfgov-refresh repository.

docker/build-dfd/build.sh

This will generate a DFD artifact named cfgov_current_build.zip in that
same root location.

I purposefully avoided using a Dockerfile here, instead choosing to pass
a complete script to a plain centos:6 container. I did this because,
without a Docker image registry or a way to periodically purge the
Docker image cache, we'd have to deal with maintaining/upgrading version
upgrades for things like pip/yarn if we included them in the base image.

[0] https://github.com/cfpb/drama-free-django