Update Wagtail admin interface for hero block
This updates the names and help text for the fields that make up the
hero block. With this update, we have incorporated designer feedback to
bring the hero module in line with the guidelines in the Design Manual,
which will make it easier for content owners to create hero graphics that
fit our site.

Note: Help text (especially links to the Design Manual from the help
text) may need to change when our new design system comes online.