Run Docker DFD build in Travis in detached mode
The drama-free-django build using a Docker container that runs in Travis
seems to fail strangely sometimes, e.g.

https://travis-ci.org/cfpb/cfgov-refresh/jobs/560138901#L1314

This works locally. I suspect this is some kind of timeout, but am not
sure. This change modifies the way that the Travis script works, so that
it runs the container in detached mode and then attaches to the logs.
I've verified that this works (at least once) on my fork.