Ensure that related_posts returns specific pages
Prior to this change `related_posts` was returning AbstractFilterPage instances. This means we can't access fields that are specific to each AbstractFilterPage subclass, like EventPage, so attempts to access the EventPage-specific datetime/fields on related items would fail.

This change adds a call to Wagtail's `specific` property on each returned page that will fetch the actual model class and fields for the page.