Fix event date range filtering
Filtering by date range on the event was using the published date of events to match the filters rather than the event dates.

The `EventArchivePage` defined a `get_form_class()` method that returned `forms.EventArchiveFilterForm` to handle this, but `get_form_class()` was never called from anywhere. This change replaces direct calls to `FilterableListForm` with calls to that method.

In addition, the specific date fields we need to match are not on the `AbstractFilterPage`, but are on `EventPage`, so this change adds a corresponding `get_model_class()` method and call to ensure that the model that's filtered can be overriden.