Ensure matching of original link string when parsing links
This change ensures that we match the original link string when we replace it by using regular expressions to extract it rather than BeautifulSoup.

Originally we were using BeautifulSoup to find all `a` tags, and then rendering the BeautifulSoup tag back to HTML and replacing that string with our marked-up replacement.

The problem is that BeautifulSoup doesn't guaratee attribute order, so when it's time to replace the original string the BeautifulSoup-rendered copy wouldn't necessary exist in the HTML string because the attribute order might've changed.

This change gets around that problem by using regular expresisons to match the original string and not engaging BeautifulSoup until it's time to add the markup.