Stop using drama-free-django empty venv
Currently drama-free-django sets up two virtual environments for each
deploy. One contains project requirements, and one is empty. To simplify
a bit, this commit stops using the empty virtualenv, because we don't
need the protection that it is intended to provide.

See mod_wsgi documentation here:

https://modwsgi.readthedocs.io/en/develop/user-guides/virtual-environments.html

The empty virtual environment is only needed and recommended "when
activating the Python virtual environment from within the WSGI script
file". We don't do that. Because we don't have multiple WSGI apps being
run within the same daemon process group, we can use a single virtual
environment.

This simplifies the configuration a bit and also makes it less specific
to drama-free-django, for example if we want to test using our Apache
config and mod_wsgi-express without using a DFD build.

To test this change, you can build a DFD artifact using
docker/drama-free-django/build.sh and then test it using the internal
Ansible process.