Stop using drama-free-django to use static.in
This project currently includes a static.in subdirectory for specifying
additional staticfiles directories for Django besides those explicitly
listed in settings. This is currently used only when using
settings.local.

In production, this same directory is configured via some behavior in
drama-free-django. To reduce dependence on drama-free-django, this
commit instead modifies settings.base so that static.in is always
included, both in local and production settings.

This change depends on https://github.com/cfpb/drama-free-django/pull/27
which allows DFD to be used while skipping its static.in behavior.