Use --no-cache for drama-free-django Docker build
The Docker-based process to build and test drama-free-django uses a
Dockerfile that includes a reference to the DFD GitHub repository.

If 'docker build' is used without '--no-cache', it's possible that
changes to that repo wouldn't get included even if the image is rebuilt.

This change adds '--no-cache' to ensure that the image is completely
rebuilt each time.