Add ability to create deployable zipfiles
This adds a new Python package to this repository that allows for the
creation of deployable zipfiles that automatically create a virtual
environment with all project requirements when they are unpacked.

This is functionality that today is provided by the drama-free-django
repository [0]; however, that project supports quite a few other
optional behaviors that are no longer needed by cf.gov. This has
motivated the migration of this deployable zipfile behavior into
cfgov-refresh.

The main interface with this new behavior is a create.py script that can
be invoked like this from the project root:

cfgov/deployable_zipfile/create.py
    ./cfgov                        # path to project code being deployed
    ./requirements/deployment.txt  # requirements to include in zipfile
    output_zipfile_name            # output filename, appended with .zip
    --extra-static=path/to/fonts   # optional, see below
    --extra-python=python3         # optional, see below

This is functionally similar to the "no-drama build" command in
drama-free-django. This script has no external (non-Python) dependencies
and thus does not need to be run inside of a virtual environment with
regular cfgov-refresh requirements.

The optional --extra-static argument can be used to include additional
static files in the zip, in a static.in subdirectory. This can be used
to store the private font files used by cf.gov, documented here [1].

Like drama-free-django, the deployable zipfile creation process involves
creating Python wheels for all requirements. This uses the version of
Python that you use to invoke the script.

You can optionally provide a second version of Python (e.g. Python 3) so
that the zip contains wheels for multiple versions. This makes it
possible to use the same zipfile to run cf.gov in Python 2 and 3. This
code brings in logic from an existing drama-free-django PR [2].

The deployable zipfile can be deployed in a way that is almost
completely compatible with how drama-free-django zipfiles work today,
like this:

path/to/deployable.zip -d /path/to/deploy/zip

If the destination directory does not exist, it will be created. To
maintain backwards compatibility, the zip is extracted to a "current"
subdirectory of the specified location.

In this subdirectory, a "venv" subdirectory is created to hold a Python
virtual environment with all requirements.

This virtual environment is configured with a special behavior related
to environment variables, consistent with drama-free-django: if an
"environment.json" file exists within the root of the deployed location
(the "current" subdirectory), it will be loaded as environment variables
on Python startup. This means that even a regular invocation of
/path/to/venv/bin/python will include those variables in the
environment.

A new Docker-based setup for generating these deployable zipfiles has
been created under docker/deployable-zipfile, comparable to the existing
docker/drama-free-django setup.

This can be invoked locally with:

docker/deployable-zipfile/build.sh

This will generate a "cfgov_current_build.zip" file in the root of this
project.

NOTE: The major difference between zipfiles generated with this approach
and those generated with drama-free-django are that DFD zips include a
wsgi.py file, and these do not. In order to use these zipfiles with our
current deployment process, you'll need to make an additional change to
our Apache wsgi.conf file here [3] to point to
${CFGOV_CURRENT}/cfgov/cfgov/wsgi.py.

[0] https://github.com/cfpb/drama-free-django
[1] https://cfpb.github.io/cfgov-refresh/installation/#webfonts
[2] drama-free-django PR 22
[3] https://github.com/cfpb/cfgov-refresh/blob/e3f0822c4787a301cceaa6eb782a25d82bcd1900/cfgov/apache/conf.d/wsgi.conf#L6