Add support for running Apache with Python 3.6
Currently the Apache configuration is hardcoded to use the Python 2.7
version of mod_wsgi. We want to optionally support running with the
Python 3.6 version instead.

Currently we install mod_wsgi using the python27-mod_wsgi package, which
installs a mod_python27-wsgi.so module. If we also install the
rh-python36-mod_wsgi package, it installs the Python 3.6-compatible
version alongside, named mod_rh-python36-wsgi.so.

In order to toggle between these two, this change relies on the presence
of a PYTHON_VERSION environment variable that needs to be passed to
Apache.