Help symlinked wsgi.py find New Relic config file
Our wsgi.py file points to a New Relic config file using a relative
path, based on our knowledge of where these two files are related to
each other in the file tree.

Using the new deployable zipfile functionality recently introduced in PR
5179, we put a symlink from a wsgi.py file in the root of the deploy
directory pointing to the real wsgi.py. This means that when wsgi.py is
accessed, its __file__ may be a symlink. This causes the relative
location of the New Relic configuration file to be wrong, which can
prevent the website from working right when NR is enabled.

This change modifies an os.path.abspath call to change it to an
os.path.realpath call, letting us always use the "real" location of
wsgi.py when locating newrelic.ini.