Fix Python 3-incompatible code in Related Posts
This change changes a use of iteritems() to items() to make the related
posts template compatible with both Python 2 and Python 3.

Currently the related posts module doesn't work in Python 3 because of
this. I've added a test case to cover this behavior.

In Python 2, going from iteritems() (an iterator) to items() (a list)
could in theory be slightly slower, but in practice the related posts
module only has a maximum of 9 items, so this shouldn't have a big
impact on performance.