Stop using symlinked wsgi.py in deployable zips
As part of the transition from drama-free-django to deployable zipfiles
generated by cfgov-refresh, we wanted to support forcing a wsgi.py in
the root of the deployed zip. To support this, the deployed zips
generated a symlink from the root to the real wsgi.py file in
cfgov/cfgov.

Now that we've transitioned, we no longer need to do this, and can have
Apache directly point to the real wsgi.py.

Because the change to the Apache config is included here along with
removal of the symlink, this won't break any old deploys and will work
going forward.