Stop use of "current" subdir in deploy zipfiles
The old drama-free-django-based deploys would always extract the release
zip into a "current" subdirectory on the target server.

For example, if a deploy zipfile were extracted by running:

/tmp/cfgov.zip -d /srv/cfgov/versions/20190828171038

then the zipfile contents would be extracted under a
/srv/cfgov/versions/20190828171038/current subdirectory, not under
/srv/cfgov/versions/20190828171038 as one might expect. This would mean,
for example, that the virtualenv was at
/srv/cfgov/versions/20190828171038/current/venv, etc.

To simplify the zipfile code, and its deployment, this change stops
extracting to a "current" subdirectory.