Add createcachetable command to initial-data.sh
This commits modifies the initial-data.sh script so that it calls the
Django createcachetable command:

https://docs.djangoproject.com/en/1.11/ref/django-admin/#createcachetable

When running either with production settings OR when one of the database
caches is enabled via environment variables, this command needs to be
called manually in order to create the necessary tables. This doesn't
happen as part of the normal migrate command.

The Apache-in-Docker setup uses production settings which enables one of
these caches, so this command is necessary for that to work right.

If the tables already exist, or if database caching isn't enabled, this
command is a no-op.