Filterable list: admin option for sorting topic filter options
Add an option to the Filterable List admin to determine how to display
the list of topics in the filter controls. None, sorted by frequency, or
sorted alphabetically. This will replace the Filter Topics checkbox in
the admin. (Will remove the Filter Topics checkbox in a future PR, soon)

Includes migrations to add the new admin field and migrate pages to use
it.