Ensure topics query string doesn't contain malicious characters
Previously, filterable lists would return a 500 error when they received
a query string with a 'topics' value containing a null character. This
would happen frequently, presumably because of bot activity.

By adding validation to the 'topics' field, instead of querying the
database with the faulty string, causing an error, we return an error to
the user saying "Topics: Select a valid choice. [your input] is not one
of the available choices".