Update paying for college models
This sets the stage for our college-cost decoder, an open version of
the dynamic disclosures tool built to help students at for-profit colleges.

Included in this PR:
- New fields on the `school` model to hold data we will collect from the College Scorecard api.
- 3 new page models to accommodate the open tool, pages presented as quizzes and static pages.
- Blocks for the quizzes.
- Initial templates and front-end resources for new pages.

Not included:
- Code to harvest the new data from Scorecard
- Final versions of templates and front-end code

This PR is intended to establish our model changes in the cfgov database so that we can
iterate on importing and using new data points and implementing a new tool desing.

move block templates to blocks directory