Handle an empty products manager
A blank search will result in `products` being a `Manager` that's empty when it's passed to `get_available_filters`. Attempting to iterate it will raise a `TypeError`. This means blank searches will 500.

This fix uses `products.all()` when iterating over products.

It also fixes the test to catch this issue. This function gets the queryset in code but was getting `objects.all()` in the test.