Initial attempt at GitHub Actions for cfgov-refresh
This change ports out Travis CI jobs to GitHub Actions as faithfully as possible. It removes the custom scripts and moves all the necessary commands and context to .github/workflows.

There's probably refactoring that could be done, and we may want to run our jobs differently on GitHub than we do on Travis, but this is intended only to migrate and get us started.

Right now it's set up with four jobs:

- `frontend`, which is a matrix to run the frontend linting and unit tests
- `backend`, which is a matrix to run the backend linting and unit tests
- `frontend_coverage`, to report code coverage from the frontend
- `backend_coverage`, to report code coverage from the backend

For code coverage, I'm trying out a utility called `diff_cover`, which reads the coverage files and then reports on the coverage difference between the branch we're on and master (this is like CodeCov's `patch` reporting). This all happens in the action (and it can happen on the command-line too!) and does not require an external service.