Allow any InfoUnitGroup format on LearnPages
Currently LearnPages can only be configured with InfoUnitGroups in the
25/75 format. This is enforced via a custom block named
InfoUnitGroup2575Only. We want to loosen these restrictions to allow for
other configurations (50/50 and 33/33/33).

This change both modifies the LearnPage to add a new InfoUnitGroup
option but also includes a data migration that should move any existing
InfoUnitGroup2575Onlys to regular InfoUnitGroup.

I'm purposefully not removing InfoUnitGroup2575Only yet to be more
conservative and roll out the change in two steps.