Use pageurl tag to reduce mega menu DB queries
This change alters the templates used to render the site mega menu so
that they use Wagtail's pageurl template tag instead of the Page.url
method to get URLs for Wagtail pages.

When using the local development setup (which disables any Django
caching used during Wagtail page lookups), this reduces the number of
database queries required to render the mega menu from 243 to 147.

To test, run a local server with the Django debug toolbar:

  ENABLE_DEBUG_TOOLBAR=1 ./runserver.sh

Visit http://localhost:8000 on master and on this branch, and compare
the number of queries. (Note that the first request when using the
Django development server will always have a few more queries due to the
way the Django contenttypes framework caches certain things.)

This change reduces database lookups because pageurl allows for the use
of the request object for caching Wagtail site root paths. Calling
Page.url from a template without providing the request object prevents
this caching, meaning that each URL lookup requires a separate site
root path lookup.

This change will have no effect in production, where a default Django
cache is configured, avoiding the need to use the request object for
this purpose.

[0] https://docs.wagtail.io/en/v1.13.4/advanced_topics/jinja2.html#pageurl
[1] https://github.com/wagtail/wagtail/blob/v1.13.4/wagtail/wagtailcore/models.py#L849